requirements.txt
git commit -m "Initial commit"mcp
crewai
crewai_tools
python-dotenv
langchain-groq
fastapi
uvicorn
litellm